Creacion evaluacion.md y test.md; y pequeñas modificaciones
