+++ Makefile
@@ -45,11 +45,9 @@ SRCS = \
 	$(BUILTINS_DIR)/export.c \
 	$(BUILTINS_DIR)/unset.c \
 	$(SIGNALS_DIR)/signals.c \
-	$(UTILS_DIR)/string_utils.c \
 	$(UTILS_DIR)/cleanup_utils.c \
 	$(UTILS_DIR)/error_utils.c \
 	$(UTILS_DIR)/utils.c \
-	$(UTILS_DIR)/sort_utils.c \
 	$(EXPANDER_DIR)/expander.c \
 	
 OBJS = $(SRC_MAIN:.c=.o) $(SRCS:.c=.o)
