+++ srcs/builtins/cd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cd.c                                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: juan-jof <juan-jof@student.42malaga.com    +#+  +:+       +#+        */
+/*   By: dponce-g <dponce-g@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/05 00:07:19 by juan-jof          #+#    #+#             */
-/*   Updated: 2025/08/15 05:06:29 by juan-jof         ###   ########.fr       */
+/*   Updated: 2025/09/17 20:11:07 by dponce-g         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,7 +61,7 @@ int	builtin_cd(char **args, t_shell *shell)
 	old_pwd = getcwd(NULL, 0);
 	if (chdir(target_dir) != 0)
 	{
-		perror("minishell: cd");
+		perror("minishell: cd: ");
 		free(old_pwd);
 		return (1);
 	}
