+++ srcs/builtins/export.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   export.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: juan-jof <juan-jof@student.42malaga.com    +#+  +:+       +#+        */
+/*   By: dponce-g <dponce-g@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/06 03:15:48 by juan-jof          #+#    #+#             */
-/*   Updated: 2025/09/16 20:26:40 by juan-jof         ###   ########.fr       */
+/*   Updated: 2025/09/17 19:48:32 by dponce-g         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
